--- qBuilder/assets/routes.js
@@ -1,88 +1,0 @@
-// These are the pages you can go to.
-// They are all wrapped in the App component, which should contain the navbar etc
-// See http://blog.mxstbr.com/2016/01/react-apps-with-pages for more information
-// about the code splitting business
-import { getHooks } from 'utils/hooks'
-
-const errorLoading = (err) => {
-  console.error('Dynamic page loading failed', err) // eslint-disable-line no-console
-}
-
-const loadModule = (cb) => (componentModule) => {
-  cb(null, componentModule.default)
-}
-
-export default function createRoutes(store) {
-  // Create reusable async injectors using getHooks factory
-  const { injectReducer } = getHooks(store)
-
-  return [
-    {
-      path: '/',
-      name: 'home',
-      getComponent(nextState, cb) {
-        const importModules = Promise.all([
-          // System.import('components/HomePage'),
-          System.import('containers/Schemas/reducer'),
-          System.import('containers/Schemas'),
-        ])
-
-        const renderRoute = loadModule(cb)
-
-        importModules.then(([reducer, component]) => {
-          injectReducer('schemas', reducer.default)
-          renderRoute(component)
-        })
-
-        importModules.catch(errorLoading)
-      },
-    }, {
-      path: '/editor',
-      name: 'new',
-      indexRoute: {
-        getComponent(nextState, cb) {
-          const importModules = Promise.all([
-            System.import('containers/Editor/reducer'),
-            System.import('containers/Editor'),
-          ])
-
-          const renderRoute = loadModule(cb)
-
-          importModules.then(([reducer, component]) => {
-            injectReducer('editor', reducer.default)
-            renderRoute(component)
-          })
-
-          importModules.catch(errorLoading)
-        }
-      },
-      childRoutes: [{
-        path: ':schemaID',
-        name: 'edit',
-        getComponent(nextState, cb) {
-          const importModules = Promise.all([
-            System.import('containers/Editor/reducer'),
-            System.import('containers/Editor'),
-          ])
-
-          const renderRoute = loadModule(cb)
-
-          importModules.then(([reducer, component]) => {
-            injectReducer('editor', reducer.default)
-            renderRoute(component)
-          })
-
-          importModules.catch(errorLoading)
-        }
-      }]
-    }, {
-      path: '*',
-      name: 'notfound',
-      getComponent(nextState, cb) {
-        System.import('components/NotFoundPage')
-          .then(loadModule(cb))
-          .catch(errorLoading)
-      },
-    },
-  ]
-}
