+++ qBuilder/assets/routes/NotFound/index.js
@@ -1,0 +1,11 @@
+export default function(injectReducer, loadModule, errorLoading) {
+  return {
+    path: '*',
+    name: 'notfound',
+    getComponent(nextState, cb) {
+      System.import('./components/NotFoundPage')
+        .then(loadModule(cb))
+        .catch(errorLoading)
+    }
+  }
+}
