+++ qBuilder/assets/routes/index.js
@@ -1,0 +1,28 @@
+// These are the pages you can go to.
+// They are all wrapped in the App component, which should contain the navbar etc
+// See http://blog.mxstbr.com/2016/01/react-apps-with-pages for more information
+// about the code splitting business
+import { getHooks } from 'utils/hooks'
+
+import SchemaList from './SchemaList'
+import Editor from './Editor'
+import NotFound from './NotFound'
+
+const errorLoading = (err) => {
+  console.error('Dynamic page loading failed', err) // eslint-disable-line no-console
+}
+
+const loadModule = (cb) => (componentModule) => {
+  cb(null, componentModule.default)
+}
+
+export default function createRoutes(store) {
+  // Create reusable async injectors using getHooks factory
+  const { injectReducer } = getHooks(store)
+
+  return [
+    SchemaList,
+    Editor,
+    NotFound
+  ].map(route => route(injectReducer, loadModule, errorLoading))
+}
